+++ Shuttle-front/package.json
@@ -1,44 +1,45 @@
 {
-  "name": "shuttle-front",
-  "version": "0.0.0",
-  "scripts": {
-    "ng": "ng",
-    "start": "ng serve",
-    "build": "ng build",
-    "watch": "ng build --watch --configuration development",
-    "test": "ng test"
-  },
-  "private": true,
-  "dependencies": {
-    "@angular/animations": "^15.0.0",
-    "@angular/cdk": "~15.0.0",
-    "@angular/common": "^15.0.0",
-    "@angular/compiler": "^15.0.0",
-    "@angular/core": "^15.0.0",
-    "@angular/forms": "^15.0.1",
-    "@angular/material": "^15.0.0",
-    "@angular/platform-browser": "^15.0.0",
-    "@angular/platform-browser-dynamic": "^15.0.0",
-    "@angular/router": "^15.0.0",
-    "@types/leaflet": "^1.9.0",
-    "@types/leaflet-routing-machine": "^3.2.4",
-    "leaflet": "^1.7.1",
-    "leaflet-routing-machine": "^3.2.12",
-    "rxjs": "~7.5.0",
-    "zone.js": "~0.12.0"
-  },
-  "devDependencies": {
-    "@angular-devkit/build-angular": "^15.0.1",
-    "@angular/cli": "~15.0.1",
-    "@angular/compiler-cli": "^15.0.0",
-    "@types/jasmine": "~4.3.0",
-    "jasmine-core": "~4.5.0",
-    "karma": "~6.4.0",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage": "~2.2.0",
-    "karma-jasmine": "~5.1.0",
-    "karma-jasmine-html-reporter": "~2.0.0",
-    "tslib": "^2.4.1",
-    "typescript": "~4.8.2"
-  }
+    "name": "shuttle-front",
+    "version": "0.0.0",
+    "scripts": {
+        "ng": "ng",
+        "start": "ng serve",
+        "build": "ng build",
+        "watch": "ng build --watch --configuration development",
+        "test": "ng test"
+    },
+    "private": true,
+    "dependencies": {
+        "@angular/animations": "^15.0.0",
+        "@angular/cdk": "~15.0.0",
+        "@angular/common": "^15.0.0",
+        "@angular/compiler": "^15.0.0",
+        "@angular/core": "^15.0.0",
+        "@angular/forms": "^15.0.1",
+        "@angular/material": "^15.0.0",
+        "@angular/platform-browser": "^15.0.0",
+        "@angular/platform-browser-dynamic": "^15.0.0",
+        "@angular/router": "^15.0.0",
+        "@types/leaflet": "^1.9.0",
+        "@types/leaflet-routing-machine": "^3.2.4",
+        "leaflet": "^1.7.1",
+        "leaflet-routing-machine": "^3.2.12",
+        "rxjs": "~7.5.0",
+        "zone.js": "~0.12.0",
+        "@auth0/angular-jwt": "^5.1.0"
+    },
+    "devDependencies": {
+        "@angular-devkit/build-angular": "^15.0.1",
+        "@angular/cli": "~15.0.1",
+        "@angular/compiler-cli": "^15.0.0",
+        "@types/jasmine": "~4.3.0",
+        "jasmine-core": "~4.5.0",
+        "karma": "~6.4.0",
+        "karma-chrome-launcher": "~3.1.0",
+        "karma-coverage": "~2.2.0",
+        "karma-jasmine": "~5.1.0",
+        "karma-jasmine-html-reporter": "~2.0.0",
+        "tslib": "^2.4.1",
+        "typescript": "~4.8.2"
+    }
 }
