+++ Shuttle-front/src/app/auth/login/component/login/login.component.ts
@@ -1,0 +1,53 @@
+import {Component} from '@angular/core';
+import {FormBuilder, FormGroup, Validators} from '@angular/forms';
+import {Router} from '@angular/router';
+import {HttpErrorResponse} from "@angular/common/http";
+import {AuthService} from "../../../auth.service";
+
+@Component({
+    selector: 'app-login',
+    templateUrl: './login.component.html',
+    styleUrls: ['./login.component.css'],
+})
+export class LoginComponent {
+    formGroup: FormGroup;
+    hasError: boolean = false;
+
+    onLoginSubmit(): void {
+        if (this.formGroup.valid) {
+            console.log(this.formGroup.getRawValue());
+            this.authService.login(this.formGroup.getRawValue()).subscribe({
+                next: (result) => {
+                    // @ts-ignore
+                    console.log(JSON.stringify(result['accessToken']));
+                    // @ts-ignore
+                    localStorage.setItem('user', JSON.stringify(result['accessToken']));
+                    this.authService.setUser();
+                    this.router.navigate(['']);
+                },
+                error: (error) => {
+                    if (error instanceof HttpErrorResponse) {
+                        this.hasError = true;
+                    }
+                },
+            });
+
+        }
+    }
+
+    constructor(private readonly formBuilder: FormBuilder, private authService: AuthService, private router: Router) {
+        this.formGroup = formBuilder.group({
+            email: ['', [Validators.required, Validators.email]],
+            password: ['', [Validators.required]],
+        });
+    }
+
+    ngOnInit() {
+        // HACK: Is there a better way to dynamically change body style from a component?
+        document.body.className = "body-gradient1"; // Defined in src/styles.css
+    }
+
+    ngOnDestroy() {
+        document.body.className = "";
+    }
+}
