+++ Shuttle-front/src/infrastructure/app-routing.module.ts
@@ -1,10 +1,10 @@
 import { NgModule } from '@angular/core';
 import { RouterModule, Routes } from '@angular/router';
 import { CreateDriverComponent } from 'src/app/admin/create-driver/create-driver.component';
-import { LoginComponent } from 'src/app/login/component/login/login.component';
-import { ForgotPasswordComponent } from 'src/app/login/component/forgot-password/forgot-password.component';
+import { LoginComponent } from 'src/app/auth/login/component/login/login.component';
+import { ForgotPasswordComponent } from 'src/app/auth/login/component/forgot-password/forgot-password.component';
 import { RegisterComponent } from 'src/app/register/register.component';
-import { ResetPasswordComponent } from 'src/app/login/component/reset-password/reset-password.component';
+import { ResetPasswordComponent } from 'src/app/auth/login/component/reset-password/reset-password.component';
 import { DriverProfileComponent } from 'src/app/driver-profile/driver-profile.component';
 import { DriverHomeComponent } from 'src/app/driver/driver-home/driver-home.component';
